pyproject.toml
[tool.poetry]
name = "django-simple-stripe"
version = "0.1.0"
description = ""
authors = ["LeandroDeJesus-S <jstleandro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
stripe = "^11.2.0"
python-dotenv = "^1.0.1"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.0.0"}

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.9.0"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q -x"
testpaths = [
    "tests",
]
DJANGO_SETTINGS_MODULE = "django_simple_stripe.settings"
python_files = ["tests/*test_*.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
